build

+++ great-wall-server/src/main/kotlin/cc/shacocloud/greatwall/model/dto/output/DurationLineMetricsOutput.kt
@@ -14,7 +14,7 @@ data class DurationLineMetricsOutput(
     /**
      * 平均值
      */
-    val avgValue: Long,
+    val avgValue: Double,
 
     /**
      * 最大值
